refactor: create src/views/Experiments subfolder
To accommodate multiple Experiments related views (help dialog)
