+++ src/router.ts
@@ -15,7 +15,7 @@ import Models from "./views/Models.vue";
 import Projects from "./views/Projects.vue";
 import ResetRequest from "./views/ResetRequest.vue";
 import PasswordReset from "./views/PasswordReset.vue";
-import Experiments from "./views/Experiments.vue";
+import Experiments from "./views/Experiments/Experiments.vue";
 import PrivacyPolicy from "./views/PrivacyPolicy.vue";
 import TermsOfService from "./views/TermsOfService.vue";
 import NotFound from "./views/NotFound.vue";
